// frontend/truckmanager/src/swagger/model/truckDTO.ts
/**
 * TruckAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-oas3
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApplicationDTO } from './applicationDTO';
import { ClassificationDTO } from './classificationDTO';
import { ColorDTO } from './colorDTO';
import { FuelTypeDTO } from './fuelTypeDTO';

export interface TruckDTO { 
    id: number;
    model: string;
    horsepower?: number;
    displacement?: number;
    fuelType?: FuelTypeDTO;
    classification?: ClassificationDTO;
    applications?: Array<ApplicationDTO>;
    colors?: Array<ColorDTO>;
    modifiedAt?: Date;
}